+++ ui/src/protos/actions/index.ts
@@ -4,7 +4,7 @@ import {Dispatch} from 'redux'
 import {client} from 'src/utils/api'
 
 // Utils
-import {addDashboardIDToCells} from 'src/dashboards/apis/v2/'
+import {addDashboardIDToCells} from 'src/dashboards/apis/'
 import {addLabelDefaults} from 'src/shared/utils/labels'
 
 // Actions
